requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alpaca-trade-api==3.2.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.0
anyio==3.7.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
cachetools==5.3.1
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.14.0
cycler==0.12.1
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.104.1
fonttools==4.61.0
frozendict==2.4.7
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jwt==1.4.0
kiwisolver==1.4.9
kombu==5.6.1
lark-parser==0.12.0
matplotlib==3.10.7
msgpack==1.0.3
multidict==6.7.0
multitasking==0.0.12
narwhals==2.13.0
networkx==3.2.1
numpy==2.3.5
numpy-financial==1.0.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
PyJWT==2.6.0
pyparsing==3.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.1
redis==5.0.1
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
seaborn==0.13.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.23
starlette==0.27.0
ta==0.10.2
trio==0.32.0
trio-websocket==0.10.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.24.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.6.4
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
yfinance==0.2.66
